# crates/lambdas/indexer-api/openapi.yaml
openapi: 3.0.3
info:
  title: Boundless Indexer API
  description: API for accessing staking, delegation, and Proof of Verifiable Work (PoVW) data for the Boundless protocol.
  version: 1.0.0
  contact:
    name: Boundless Development Team
servers:
  - url: /
    description: Current server
...
paths:
...
  /v1/povw:
    get:
      tags:
        - PoVW
      summary: Get PoVW Summary Statistics
      description: Get global PoVW summary statistics across all epochs
      operationId: getPovwSummary
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PovwAggregateResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/povw/addresses:
    get:
      tags:
        - PoVW
      summary: Get All-Time PoVW Leaderboard
      description: Get the all-time PoVW leaderboard ranking all work log IDs by total rewards
      operationId: getPovwLeaderboard
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PovwAggregateResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/povw/epochs:
    get:
      tags:
        - PoVW
      summary: Get All Epochs Summary
      description: Get a list of all epochs with summary statistics for each
      operationId: getAllEpochsSummary
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpochsSummaryResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/povw/epochs/{epoch}:
    get:
      tags:
        - PoVW
      summary: Get Epoch Summary
      description: Get summary statistics for a specific epoch
      operationId: getPovwByEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PovwByEpochResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/povw/epochs/{epoch}/addresses:
    get:
      tags:
        - PoVW
      summary: Get Epoch PoVW Leaderboard
      description: Get the PoVW leaderboard for a specific epoch
      operationId: getEpochPovwLeaderboard
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PovwByEpochResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/povw/addresses/{address}:
    get:
      tags:
        - PoVW
      summary: Get PoVW History by Address
      description: Get complete PoVW rewards history for a specific work log ID across all epochs
      operationId: getPovwHistoryByAddress
      parameters:
        - $ref: '#/components/parameters/AddressParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: start_epoch
          in: query
          description: Filter results from this epoch onwards
          schema:
            type: integer
            minimum: 0
        - name: end_epoch
          in: query
          description: Filter results up to this epoch
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PovwByEpochResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/povw/epochs/{epoch}/addresses/{address}:
    get:
      tags:
        - PoVW
      summary: Get PoVW by Address and Epoch
      description: Get PoVW rewards for a specific work log ID at a specific epoch
      operationId: getPovwByAddressAndEpoch
      parameters:
        - $ref: '#/components/parameters/EpochParam'
        - $ref: '#/components/parameters/AddressParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PovwRewardByEpoch'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    OffsetParam:
      name: offset
      in: query
      description: Number of results to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
    EpochParam:
      name: epoch
      in: path
      required: true
      description: Epoch number (2-day period)
      schema:
        type: integer
        minimum: 0
    AddressParam:
      name: address
      in: path
      required: true
      description: Ethereum address (0x-prefixed, checksummed)
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$'
        example: '0x2408e37489c231f883126c87e8aadbad782a040a'
  schemas:
    Pagination:
      type: object
      required:
        - count
        - offset
        - limit
      properties:
        count:
          type: integer
          description: Number of items in this response
        offset:
          type: integer
          description: Offset used for this response
        limit:
          type: integer
          description: Maximum items per response
...
    PovwRewardAggregate:
      type: object
      required:
        - work_log_id
        - total_work_submitted
        - total_actual_rewards
        - total_uncapped_rewards
        - epochs_participated
      properties:
        rank:
          type: integer
          nullable: true
          description: Leaderboard rank
        work_log_id:
          type: string
          description: Work log ID (Ethereum address format)
        total_work_submitted:
          type: string
          description: Total work submitted across all epochs
        total_actual_rewards:
          type: string
          description: Total actual rewards received (after capping)
        total_uncapped_rewards:
          type: string
          description: Total rewards before capping
        epochs_participated:
          type: integer
          description: Number of epochs participated
    PovwRewardByEpoch:
      type: object
      required:
        - work_log_id
        - epoch
        - work_submitted
        - percentage
        - uncapped_rewards
        - reward_cap
        - actual_rewards
        - is_capped
        - staked_amount
      properties:
        rank:
          type: integer
          nullable: true
          description: Leaderboard rank (only in list responses)
        work_log_id:
          type: string
          description: Work log ID (Ethereum address format)
        epoch:
          type: integer
          description: Epoch number
        work_submitted:
          type: string
          description: Amount of work submitted in this epoch
        percentage:
          type: number
          format: double
          description: Percentage of total work in epoch
        uncapped_rewards:
          type: string
          description: Rewards before capping (in wei)
        reward_cap:
          type: string
          description: Maximum rewards allowed based on stake (in wei)
        actual_rewards:
          type: string
          description: Actual rewards received (in wei)
        is_capped:
          type: boolean
          description: Whether rewards were capped
        staked_amount:
          type: string
          description: Amount staked (in wei)
...
    PovwAggregateResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PovwRewardAggregate'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PovwByEpochResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/PovwRewardByEpoch'
        pagination:
          $ref: '#/components/schemas/Pagination'
    EpochsSummaryResponse:
      type: object
      required:
        - entries
        - pagination
      properties:
        entries:
          type: array
          items:
            $ref: '#/components/schemas/EpochPoVWSummary'
        pagination:
          $ref: '#/components/schemas/Pagination'
    EpochPoVWSummary:
      type: object
      required:
        - epoch
        - total_work
        - total_work_formatted
        - total_emissions
        - total_emissions_formatted
        - total_capped_rewards
        - total_capped_rewards_formatted
        - total_uncapped_rewards
        - total_uncapped_rewards_formatted
        - epoch_start_time
        - epoch_end_time
        - num_participants
      properties:
        epoch:
          type: integer
          description: Epoch number
        total_work:
          type: string
          description: Total work submitted in this epoch
        total_work_formatted:
          type: string
          description: Total work (human-readable)
        total_emissions:
          type: string
          description: Total emissions for this epoch (in wei)
        total_emissions_formatted:
          type: string
          description: Total emissions (human-readable)
        total_capped_rewards:
          type: string
          description: Total capped rewards distributed (in wei)
        total_capped_rewards_formatted:
          type: string
          description: Total capped rewards (human-readable)
        total_uncapped_rewards:
          type: string
          description: Total uncapped rewards before capping (in wei)
        total_uncapped_rewards_formatted:
          type: string
          description: Total uncapped rewards (human-readable)
        epoch_start_time:
          type: integer
          description: Unix timestamp when epoch started
        epoch_end_time:
          type: integer
          description: Unix timestamp when epoch ends
        num_participants:
          type: integer
          description: Number of participants in this epoch
        last_updated_at:
          type: string
          nullable: true
          description: Last update timestamp
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message
  responses:
    BadRequest:
      description: Bad Request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Bad Request
            message: Invalid address format
    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not Found
            message: No data found for the specified parameters
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal Server Error
            message: An unexpected error occurred